shape_checking_study/shape_guard/pyproject.toml
[tool.poetry]
name = "shape_checking_study"
version = "0.1.0"
description = ""
authors = ["Jesper Nielsen <jesper@prowler.io>"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "artifactory"
url = "https://artifactory.piointernal.prowler.io/artifactory/api/pypi/pypi-prod/simple"
default = true

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.10.1"
numpy = "1.19.0"
torch-shapeguard = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^20.0b0"
isort = "^5.10.1"
mypy = "^0.920"
pylint = "^2.10.0"
pytest = "^5.2"
taskipy = "^1.9.0"

[tool.black]
line-length = 95
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 95

[tool.taskipy.tasks]

isort = "isort shape_checking_study tests"
black = "black shape_checking_study tests"
mypy = "mypy shape_checking_study tests"
pylint = "pylint shape_checking_study tests"
pytest = "pytest tests"
pytorch_example = "python -m shape_checking_study.pytorch_example"

format = """\
  echo XXXXXXXXXXXXXXXXXXXXXXXXX isort XXXXXXXXXXXXXXXXXXXXXXXXX ; \
  task isort ; \
  echo XXXXXXXXXXXXXXXXXXXXXXXXX black XXXXXXXXXXXXXXXXXXXXXXXXX ; \
  task black ; \
"""

lint = """\
  echo XXXXXXXXXXXXXXXXXXXXXXXXX mypy XXXXXXXXXXXXXXXXXXXXXXXXX ; \
  task mypy ;
  echo XXXXXXXXXXXXXXXXXXXXXXXXX pylint XXXXXXXXXXXXXXXXXXXXXXXXX ; \
  task pylint ; \
"""

benchmark = """\
  task format ; \
  task lint ; \
  echo XXXXXXXXXXXXXXXXXXXXXXXXX pytorch XXXXXXXXXXXXXXXXXXXXXXXXX ; \
  task pytorch_example ; \
"""

test = """\
  task format ; \
  task lint ; \
  echo XXXXXXXXXXXXXXXXXXXXXXXXX pytest XXXXXXXXXXXXXXXXXXXXXXXXX ; \
  task pytest ; \
"""
